requirements.txt
llama-index
llama-index-core
llama-index-llms-openai
llama-index-embeddings-openai
llama-index-vector-stores-qdrant
streamlit
python-dotenv
sentence-transformers
qdrant-client
mcp